Renamed defualt import export tab component
